Enable additional compile-time checking.
Besides some makefile twiddling to match changes in libos,
this fixes many of the issues found by sparse and the other
additional warnings.

Many of them were harmless (but still wrong), some were a product of the
annoying lack of a way in C to declare an integer whose range is a subset
of "int" and "unsigned int", so that its signedness can be interpreted in
the context of its expression -- but a few were real bugs (e.g.
parameters not passed properly due to not getting a prototype from a
header) or had a high probability of turning into a bug (e.g. local
variables shadowing global variables).

The GDB stub had many instances of inline functions being declared
prior to definition, which sparse doesn't like.  I removed all instances
of inline in that file (let the compiler decide when to inline), and
the size of the hypervisor shrunk by 18KB. :-)

Please don't go inline-crazy.

+++ src/hcalls.c
@@ -283,8 +283,8 @@ static void fh_byte_channel_send(trapframe_t *regs)
 		regs->gpregs[3] = EAGAIN;
 	else {
 		/* put chars into bytechannel queue here */
-		int ret = byte_chan_send(bc, buf, len);
-		assert(ret == len);
+		ssize_t ret = byte_chan_send(bc, buf, len);
+		assert(ret == (ssize_t)len);
 		regs->gpregs[3] = 0;  /* success */
 	}
